requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
antiorm==1.2.1
anytree==2.8.0
appdirs==1.4.4
APScheduler==3.7.0
asgiref==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
Backtesting==0.3.1
beautifulsoup4==4.9.3
binance-futures==1.1.0
binance.py==1.8.0
bleach==3.3.0
bokeh==2.3.1
bs4==0.0.1
bullet==2.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cliff==3.7.0
cloudpickle==1.6.0
cmd2==1.5.0
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
cycler==0.10.0
dateparser==1.0.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
distlib==0.3.1
Django==3.2.4
djangorestframework==3.12.4
dnspython==2.0.0
docutils==0.17.1
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
gnupg==2.3.1
gunicorn==20.1.0
html5lib==1.1
humanfriendly==9.1
idna==2.10
imagesize==1.2.0
importlib-metadata==4.0.1
ipwhois==1.2.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
keyboard==0.13.5
keyring==23.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
MouseInfo==0.1.3
mpl-finance==0.10.1
mplfinance==0.12.7a17
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
pluggy==0.13.1
prettytable==2.1.0
prompt-toolkit==3.0.18
property-manager==3.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
PyAutoGUI==0.9.52
PyAutoIt==0.6.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyGetWindow==0.0.9
Pygments==2.9.0
pyhomebroker @ git+https://github.com/crapher/pyhomebroker@f8161339f4186ec54dc304a3626725267d9b63dc
pymatriz==0.1.4
PyMsgBox==1.0.9
pynput==1.7.3
pyparsing==2.4.7
pyperclip==1.8.2
pyquery==1.4.3
pyreadline==2.1
pyreadline3==3.3
PyRect==0.1.4
PyScreeze==0.1.27
python-binance==1.0.10
python-dateutil==2.8.1
python-dotenv==0.19.0
python-secrets==21.2.0
PyTweening==1.0.3
pytz==2021.1
pywhatkit==4.7
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.0.3
readme-renderer==29.0
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
signalr-client-threads==0.0.12
simplejson==3.17.5
six==1.15.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder-kernels==2.0.3
sqlparse==0.4.1
sseclient==0.0.27
stevedore==3.3.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tox==3.23.1
tqdm==4.60.0
traitlets==5.0.5
twine==3.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
verboselogs==1.7
virtualenv==20.4.6
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
websocket-client==0.59.0
websockets==9.0.1
Werkzeug==2.0.1
wikipedia==1.4.0
xkcdpass==1.19.2
xlrd==2.0.1
yamlreader==3.0.4
yarl==1.6.3
yfinance==0.1.59
zipimportx==0.3.2
zipp==3.4.1
